SHUTTER_0 U1b — last_capture_s is a double, because PLUMB_0 B1 says so
DEVIATION FROM VERBATIM, in its own commit so it can be dropped alone.

The handoff declares `float last_capture_s = -1.0e9f` and stamps it `(float)now`.
It is differenced against TimeState::seconds, which is the exact pattern PLUMB_0 B1
governs, and every other such stamp in the tree is a double — spawn_engine's
lastCensusDump_, cartridge's rosterGolResidueDump_ and lastCardTick_, floaters'
last_alloc_time, and defer_since two lines below it in this very struct, whose
comment reads "PLUMB_0 B1 — differenced against TimeState::seconds". This was the
only float one.

Not the float-accumulator stall PLUMB_0 measured (this is a stamp, not an
accumulator) but the same family: a float stamp quantises at the wall clock's
magnitude. Measured ulp against the 5 s floor — 0.031 s at four days (PLUMB_0's own
"four days is not a hypothetical"), 0.25 s at thirty days, 1.0 s at six months,
4.0 s at 485 days, where the floor stops being measurable. The piece is permanently
hosted.

Nothing else changes: same floor, same ordering, same behaviour on any session short
enough for float to have been fine.

+++ src/cartridges/the_board/bodies/gallery.hpp
@@ -713,7 +713,17 @@ struct PhotographerState {
     // SHUTTER_0 — when the last capture FIRED, in TimeState seconds; the
     // spacing floor measures from here. Very negative: the first shot of a
     // session owes no spacing.
-    float last_capture_s = -1.0e9f;
+    //
+    // DOUBLE, BY PLUMB_0 B1, LIKE defer_since BELOW AND EVERY OTHER STAMP
+    // IN THE TREE THAT IS DIFFERENCED AGAINST TimeState::seconds. The
+    // handoff wrote float. A float STAMP does not stall the way PLUMB_0's
+    // float ACCUMULATOR did, but it quantises: the ulp of a float at the
+    // wall clock's magnitude is 0.031 s at four days (PLUMB_0's own "not a
+    // hypothetical" horizon), 0.25 s at thirty, 1.0 s at six months and
+    // 4.0 s at 485 days — at which point a 5-second floor cannot be
+    // measured at all. The piece is permanently hosted. One word, and the
+    // law that already governs the member two lines down governs this one.
+    double last_capture_s = -1.0e9;
     float next_threshold = PhotographerCaptureConfig::TRIGGER_DISTANCE_MEAN;
     uint32_t pending_shots = 0;
     float prev_point_x = 0.0f;
@@ -1681,7 +1691,7 @@ inline void update_photographer(GalleryState& gs, GalleryDeps* c, wgpu::Queue& q
 
         gs.photographer.defer_since = -1.0f;
         capture_snapshot(gs, c, px, pz, queue);
-        gs.photographer.last_capture_s = (float)now;   // SHUTTER_0 — the floor measures from the fire
+        gs.photographer.last_capture_s = now;   // SHUTTER_0 — the floor measures from the fire (double: PLUMB_0 B1)
         gs.photographer.pending_shots--;
         gs.photographer.frame_cooldown = PhotographerCaptureConfig::BURST_COOLDOWN_FRAMES;
         return;
